+++ lib/features/purchase/data/datasources/purchase_remote_data_source.dart
@@ -1,12 +1,15 @@
 import 'package:coffeecard/core/errors/failures.dart';
+import 'package:coffeecard/core/extensions/either_extensions.dart';
 import 'package:coffeecard/core/network/network_request_executor.dart';
+import 'package:coffeecard/features/purchase/data/models/initiate_purchase_model.dart';
+import 'package:coffeecard/features/purchase/data/models/single_purchase_model.dart';
+import 'package:coffeecard/features/purchase/domain/entities/initiate_purchase.dart';
+import 'package:coffeecard/features/purchase/domain/entities/single_purchase.dart';
 import 'package:coffeecard/generated/api/coffeecard_api_v2.swagger.dart';
-import 'package:coffeecard/models/purchase/initiate_purchase.dart';
-import 'package:coffeecard/models/purchase/single_purchase.dart';
 import 'package:dartz/dartz.dart';
 
-class PurchaseRepository {
-  PurchaseRepository({
+class PurchaseRemoteDataSource {
+  PurchaseRemoteDataSource({
     required this.apiV2,
     required this.executor,
   });
@@ -20,26 +23,22 @@ class PurchaseRepository {
     int productId,
     PaymentType paymentType,
   ) async {
-    final result = await executor(
+    return executor(
       () => apiV2.apiV2PurchasesPost(
         body: InitiatePurchaseRequest(
           productId: productId,
           paymentType: paymentTypeToJson(paymentType),
         ),
       ),
-    );
-
-    return result.map(InitiatePurchase.fromDto);
+    ).bindFuture(InitiatePurchaseModel.fromDto);
   }
 
   /// Get a purchase by its purchase id
   Future<Either<NetworkFailure, SinglePurchase>> getPurchase(
     int purchaseId,
   ) async {
-    final result = await executor(
+    return executor(
       () => apiV2.apiV2PurchasesIdGet(id: purchaseId),
-    );
-
-    return result.map(SinglePurchase.fromDto);
+    ).bindFuture(SinglePurchaseModel.fromDto);
   }
 }
