+++ lib/features/purchase/domain/entities/internal_payment_type.dart
@@ -1,0 +1,4 @@
+enum InternalPaymentType {
+  mobilePay,
+  free,
+}
