+++ lib/features/purchase/presentation/pages/buy_tickets_page.dart
@@ -1,11 +1,14 @@
 import 'package:coffeecard/base/strings.dart';
 import 'package:coffeecard/base/style/colors.dart';
+import 'package:coffeecard/cubits/environment/environment_cubit.dart';
 import 'package:coffeecard/cubits/products/products_cubit.dart';
 import 'package:coffeecard/data/repositories/v1/product_repository.dart';
+import 'package:coffeecard/features/purchase/domain/entities/payment.dart';
+import 'package:coffeecard/features/purchase/domain/entities/payment_status.dart';
 import 'package:coffeecard/features/receipt/presentation/cubit/receipt_cubit.dart';
+import 'package:coffeecard/features/receipt/presentation/widgets/receipt_overlay.dart';
 import 'package:coffeecard/features/ticket/presentation/cubit/tickets_cubit.dart';
-import 'package:coffeecard/models/purchase/payment.dart';
-import 'package:coffeecard/models/purchase/payment_status.dart';
+import 'package:coffeecard/models/environment.dart';
 import 'package:coffeecard/models/ticket/product.dart';
 import 'package:coffeecard/service_locator.dart';
 import 'package:coffeecard/utils/firebase_analytics_event_logging.dart';
@@ -18,14 +21,14 @@ import 'package:coffeecard/widgets/components/tickets/buy_tickets_card.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-class BuySingleDrinkPage extends StatelessWidget {
-  const BuySingleDrinkPage();
+class BuyTicketsPage extends StatelessWidget {
+  const BuyTicketsPage();
 
-  static const String _fbAnalyticsListId = 'buy_one_drink';
-  static const String _fbAnalyticsListName = Strings.buyOneDrink;
+  static const String _fbAnalyticsListId = 'buy_tickets';
+  static const String _fbAnalyticsListName = Strings.buyTickets;
 
   static Route get route =>
-      MaterialPageRoute(builder: (_) => const BuySingleDrinkPage());
+      MaterialPageRoute(builder: (_) => const BuyTicketsPage());
 
   @override
   Widget build(BuildContext context) {
@@ -33,14 +36,14 @@ class BuySingleDrinkPage extends StatelessWidget {
       create: (context) =>
           ProductsCubit(sl.get<ProductRepository>())..getProducts(),
       child: AppScaffold.withTitle(
-        title: Strings.buyOneDrink,
+        title: Strings.buyTickets,
         body: BlocBuilder<ProductsCubit, ProductsState>(
           builder: (context, state) {
             if (state is ProductsLoading) {
               return const Loading(loading: true);
             } else if (state is ProductsLoaded) {
               sl<FirebaseAnalyticsEventLogging>().viewProductsListEvent(
-                state.singleDrinkProducts,
+                state.ticketProducts,
                 _fbAnalyticsListId,
                 _fbAnalyticsListName,
               );
@@ -51,11 +54,11 @@ class BuySingleDrinkPage extends StatelessWidget {
                   gap: GridGap.normal,
                   gapSmall: GridGap.tight,
                   singleColumnOnSmallDevice: true,
-                  children: state.singleDrinkProducts
+                  children: state.ticketProducts
                       .map(
                         (product) => BuyTicketsCard(
                           product: product,
-                          onTap: buyNSwipeModal,
+                          onTap: buyTicketsModal,
                         ),
                       )
                       .toList(),
@@ -63,6 +66,7 @@ class BuySingleDrinkPage extends StatelessWidget {
               );
             } else if (state is ProductsError) {
               return ErrorSection(
+                center: true,
                 error: state.error,
                 retry: context.read<ProductsCubit>().getProducts,
               );
@@ -75,7 +79,7 @@ class BuySingleDrinkPage extends StatelessWidget {
     );
   }
 
-  Future<void> buyNSwipeModal(
+  Future<void> buyTicketsModal(
     BuildContext context,
     Product product,
     State state,
@@ -97,7 +101,7 @@ class BuySingleDrinkPage extends StatelessWidget {
         useRootNavigator: true,
         builder: (_) => BuyTicketBottomModalSheet(
           product: product,
-          description: Strings.paymentConfirmationTopSingle(
+          description: Strings.paymentConfirmationTopTickets(
             product.amount,
             product.name,
           ),
@@ -105,13 +109,24 @@ class BuySingleDrinkPage extends StatelessWidget {
       );
       if (!state.mounted) return;
       if (payment != null && payment.status == PaymentStatus.completed) {
-        // Send the user back to the home-screen
+        // Send the user back to the home-screen.
         Navigator.pop(context);
 
-        final ticketCubit = context.read<TicketsCubit>();
-        final receiptCubit = context.read<ReceiptCubit>();
-        await ticketCubit.useTicket(product.id);
-        await receiptCubit.fetchReceipts();
+        final envState = context.read<EnvironmentCubit>().state;
+
+        final updateTicketsRequest = context.read<TicketsCubit>().getTickets();
+        final updateReceiptsRequest =
+            context.read<ReceiptCubit>().fetchReceipts();
+
+        ReceiptOverlay.of(context).show(
+          isTestEnvironment:
+              envState is EnvironmentLoaded && envState.env.isTest,
+          isPurchase: true,
+          productName: payment.productName,
+          timeUsed: payment.purchaseTime,
+        );
+        await updateTicketsRequest;
+        await updateReceiptsRequest;
       }
     }
   }
