+++ lib/features/purchase/presentation/widgets/purchase_overlay.dart
@@ -1,9 +1,12 @@
 import 'package:coffeecard/base/style/colors.dart';
-import 'package:coffeecard/cubits/purchase/purchase_cubit.dart';
-import 'package:coffeecard/models/purchase/payment.dart';
+import 'package:coffeecard/features/purchase/data/repositories/payment_handler.dart';
+import 'package:coffeecard/features/purchase/domain/entities/internal_payment_type.dart';
+import 'package:coffeecard/features/purchase/domain/entities/payment.dart';
+import 'package:coffeecard/features/purchase/domain/usecases/init_purchase.dart';
+import 'package:coffeecard/features/purchase/domain/usecases/verify_purchase_status.dart';
+import 'package:coffeecard/features/purchase/presentation/cubit/purchase_cubit.dart';
+import 'package:coffeecard/features/purchase/presentation/widgets/purchase_process.dart';
 import 'package:coffeecard/models/ticket/product.dart';
-import 'package:coffeecard/payment/payment_handler.dart';
-import 'package:coffeecard/widgets/components/purchase/purchase_process.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -21,11 +24,17 @@ Future<Payment?> showPurchaseOverlay({
           // Will prevent Android back button from closing overlay.
           onWillPop: () async => false,
           child: BlocProvider(
-            create: (context) => PurchaseCubit(
-              paymentHandler:
-                  PaymentHandler.createPaymentHandler(paymentType, context),
-              product: product,
-            ),
+            create: (context) {
+              final paymentHandler =
+                  PaymentHandler.createPaymentHandler(paymentType, context);
+
+              return PurchaseCubit(
+                product: product,
+                initPurchase: InitPurchase(paymentHandler: paymentHandler),
+                verifyPurchaseStatus:
+                    VerifyPurchaseStatus(paymentHandler: paymentHandler),
+              );
+            },
             child: BlocListener<PurchaseCubit, PurchaseState>(
               listener: (context, state) async {
                 if (state is PurchaseCompleted) {
