+++ lib/features/user/domain/entities/role.dart
@@ -5,10 +5,8 @@ enum Role {
   customer,
   barista,
   manager,
-  board,
-}
+  board;
 
-extension RoleExtension on Role {
   // json is a dynamic object by its very nature
   //ignore: avoid-dynamic
   static Role fromJson(dynamic json) {
